Rename timetables folders and component
