+++ ui/src/components/timetables/index.ts
@@ -1,0 +1,2 @@
+export * from './main';
+export * from './vehicle-schedule-details';
