+++ ui/src/components/timetables/main/index.ts
@@ -1,0 +1,1 @@
+export * from './TimetablesMainPage';
