+++ ui/src/components/timetables/passing-times-by-stop/PassingTimesByStopTable.tsx
