+++ ui/src/components/timetables/passing-times-by-stop/PassingTimesByStopTableRow.tsx
