+++ ui/src/components/timetables/passing-times-by-stop/PassingTimesByStopTableRowPassingMinute.tsx
