+++ ui/src/components/timetables/passing-times-by-stop/PassingTimesByStopTableRowPassingTime.tsx
