+++ ui/src/components/timetables/passing-times-by-stop/VehicleJourneyPopover.tsx
