+++ ui/src/components/timetables/passing-times-by-stop/index.ts
