+++ ui/src/components/timetables/vehicle-schedule-details/index.ts
@@ -1,0 +1,2 @@
+export * from './vehicle-service-table';
+export * from './VehicleScheduleDetailsPage';
