+++ ui/src/components/timetables/vehicle-schedule-details/vehicle-service-table/VehicleServiceTableRow.tsx
@@ -1,4 +1,4 @@
-import { padToTwoDigits } from '../../../time';
+import { padToTwoDigits } from '../../../../time';
 
 interface Props {
   data: VehicleServiceRowData;
