+++ ui/src/components/timetables/vehicle-schedule-details/vehicle-service-table/index.ts
