--- ui/src/components/vehicle-service/index.ts
@@ -1,3 +1,0 @@
-export * from './VehicleScheduleDetailsPage';
-export * from './VehicleScheduleListPage';
-export * from './vehicle-service-table';
